repository 모델로 ImageFragment 적용
