+++ doubleclutch30/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/data/repository/NaverSearchRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package com.example.architecturestudy.data.repository
 
 import com.example.architecturestudy.data.model.BlogData
+import com.example.architecturestudy.data.model.ImageData
 import com.example.architecturestudy.data.model.KinData
 import com.example.architecturestudy.data.model.MovieData
 import com.example.architecturestudy.data.source.remote.NaverSearchRemoteDataSource
@@ -30,6 +31,14 @@ class NaverSearchRepositoryImpl : NaverSearchRepository {
 
     override fun getKin(keyword: String, success: (KinData) -> Unit, fail: (Throwable) -> Unit) {
         naverSearchRemoteDataSource.getKin(
+            keyword = keyword,
+            success = success,
+            fail = fail
+        )
+    }
+
+    override fun getImage(keyword: String, success: (ImageData) -> Unit, fail: (Throwable) -> Unit) {
+        naverSearchRemoteDataSource.getImage(
             keyword = keyword,
             success = success,
             fail = fail
