+++ doubleclutch30/ArchitectureStudy/app/src/main/java/com/example/architecturestudy/ui/image/ImageFragment.kt
@@ -9,6 +9,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.architecturestudy.R
 import com.example.architecturestudy.data.model.ImageData
+import com.example.architecturestudy.data.repository.NaverSearchRepositoryImpl
 import com.example.architecturestudy.network.Api
 import com.example.architecturestudy.network.ApiClient
 import kotlinx.android.synthetic.main.fragment_image.*
@@ -22,17 +23,28 @@ class ImageFragment : Fragment() {
 
     private lateinit var imageAdapter: ImageAdapter
 
+    private val naverSearchRepository by lazy { NaverSearchRepositoryImpl() }
+
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        imageAdapter = ImageAdapter()
         return inflater.inflate(R.layout.fragment_image, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
+
+        imageAdapter = ImageAdapter()
+
+        recycleview.apply {
+            adapter = imageAdapter
+            layoutManager = LinearLayoutManager(activity)
+            addItemDecoration(DividerItemDecoration(activity, DividerItemDecoration.VERTICAL))
+        }
+
         btn_search.setOnClickListener {
             if(input_text != null) {
                 val edit = edit_text.text.toString()
@@ -42,20 +54,25 @@ class ImageFragment : Fragment() {
     }
 
     private fun searchImageList(keyword : String) {
-        restClient.requestImage(keyword).enqueue(object : Callback<ImageData> {
-            override fun onFailure(call: Call<ImageData>, t: Throwable) {
-                error(message = t.toString())
-            }
-
-            override fun onResponse(call: Call<ImageData>, response: Response<ImageData>) {
-                if(response.isSuccessful) {
-                    response.body()?.items?.let {
-                        imageListAdapter()
-                        imageAdapter.update(it)
-                    }
-                }
-            }
-        })
+        naverSearchRepository.getImage(
+            keyword = keyword,
+            success = { responsImage -> imageAdapter.update(responsImage.items) },
+            fail = { e -> error(message = e.toString()) }
+        )
+//        restClient.requestImage(keyword).enqueue(object : Callback<ImageData> {
+//            override fun onFailure(call: Call<ImageData>, t: Throwable) {
+//                error(message = t.toString())
+//            }
+//
+//            override fun onResponse(call: Call<ImageData>, response: Response<ImageData>) {
+//                if(response.isSuccessful) {
+//                    response.body()?.items?.let {
+//                        imageListAdapter()
+//                        imageAdapter.update(it)
+//                    }
+//                }
+//            }
+//        })
     }
 
     private fun imageListAdapter() {
